Fix Repa and Dph benchmarks
The benchmarks were not compiling correctly. There were two
issues. First, the using only the -mem2reg option was
causing the ll files to get a validation error. We also
needed to pass the -constprop option for these benchmarks.
We added the --extra-opt-pre-link argument to htrace to
specify these extra optimization passes.

Second, the benchmarks needed to link with libffi to run
correctly. We fixed this by specifying an ffi dependency in
the fibon.extra.cfg file.